fix: Upload files at any folder
